use a single uint256 as encoded swap

+++ scripts/templates/MesonConfig.sol
@@ -9,7 +9,7 @@ contract MesonConfig {
   // Ref https://github.com/satoshilabs/slips/blob/master/slip-0044.md
   bytes4 constant COIN_TYPE = CONFIG_COIN_TYPE;
 
-  uint48 constant MIN_BOND_TIME_PERIOD = 1 hours;
-  uint48 constant MAX_BOND_TIME_PERIOD = 2 hours;
-  uint48 constant LOCK_TIME_PERIOD = 20 minutes;
+  uint40 constant MIN_BOND_TIME_PERIOD = 1 hours;
+  uint40 constant MAX_BOND_TIME_PERIOD = 2 hours;
+  uint40 constant LOCK_TIME_PERIOD = 20 minutes;
 }
